+++ wpilibc/athena/include/Counter.h
@@ -77,7 +77,7 @@ class Counter : public SensorBase,
   void SetUpDownCounterMode();
   void SetExternalDirectionMode();
   void SetSemiPeriodMode(bool highSemiPeriod);
-  void SetPulseLengthMode(float threshold);
+  void SetPulseLengthMode(double threshold);
 
   void SetReverseDirection(bool reverseDirection);
