+++ wpilibc/athena/src/Encoder.cpp
@@ -296,10 +296,10 @@ bool Encoder::GetDirection() const {
  * The scale needed to convert a raw counter value into a number of encoder
  * pulses.
  */
-float Encoder::DecodingScaleFactor() const {
+double Encoder::DecodingScaleFactor() const {
   if (StatusIsFatal()) return 0.0;
   int32_t status = 0;
-  float val = HAL_GetEncoderDecodingScaleFactor(m_encoder, &status);
+  double val = HAL_GetEncoderDecodingScaleFactor(m_encoder, &status);
   wpi_setErrorWithContext(status, HAL_GetErrorMessage(status));
   return val;
 }
@@ -310,10 +310,10 @@ float Encoder::DecodingScaleFactor() const {
  * @return The distance driven since the last reset as scaled by the value from
  *         SetDistancePerPulse().
  */
-float Encoder::GetDistance() const {
+double Encoder::GetDistance() const {
   if (StatusIsFatal()) return 0.0;
   int32_t status = 0;
-  float value = HAL_GetEncoderDistance(m_encoder, &status);
+  double value = HAL_GetEncoderDistance(m_encoder, &status);
   wpi_setErrorWithContext(status, HAL_GetErrorMessage(status));
   return value;
 }
@@ -326,10 +326,10 @@ float Encoder::GetDistance() const {
  *
  * @return The current rate of the encoder.
  */
-float Encoder::GetRate() const {
+double Encoder::GetRate() const {
   if (StatusIsFatal()) return 0.0;
   int32_t status = 0;
-  float value = HAL_GetEncoderRate(m_encoder, &status);
+  double value = HAL_GetEncoderRate(m_encoder, &status);
   wpi_setErrorWithContext(status, HAL_GetErrorMessage(status));
   return value;
 }
@@ -340,7 +340,7 @@ float Encoder::GetRate() const {
  * @param minRate The minimum rate.  The units are in distance per second as
  *                scaled by the value from SetDistancePerPulse().
  */
-void Encoder::SetMinRate(float minRate) {
+void Encoder::SetMinRate(double minRate) {
   if (StatusIsFatal()) return;
   int32_t status = 0;
   HAL_SetEncoderMinRate(m_encoder, minRate, &status);
@@ -364,7 +364,7 @@ void Encoder::SetMinRate(float minRate) {
  * @param distancePerPulse The scale factor that will be used to convert pulses
  *                         to useful units.
  */
-void Encoder::SetDistancePerPulse(float distancePerPulse) {
+void Encoder::SetDistancePerPulse(double distancePerPulse) {
   if (StatusIsFatal()) return;
   int32_t status = 0;
   HAL_SetEncoderDistancePerPulse(m_encoder, distancePerPulse, &status);
